+++ agent/src/ebpf/user/extended/extended.c
@@ -36,6 +36,9 @@ int __attribute__ ((weak)) extended_maps_set(struct bpf_tracer *tracer)
 	return 0;
 }
 
+void __attribute__ ((weak)) extended_prog_jump_tables(struct bpf_tracer *tracer) {
+}
+
 int __attribute__ ((weak)) extended_proc_event_handler(int pid,
 						       const char *name,
 						       enum proc_act_type type)
@@ -55,3 +58,8 @@ void __attribute__ ((weak)) extended_events_handle(void) {}
 void __attribute__ ((weak)) extended_process_exit(int pid) {}
 
 void __attribute__ ((weak)) extended_match_pid_handle(int feat, int pid, enum match_pids_act act) {}
+
+bool __attribute__ ((weak)) extended_require_dwarf(int pid, const char *name)
+{
+	return false;
+}